implemented 'Text -> Builder' UTF-8 encoders
It uses a coupled end-of-input-and-output boundary and exploits the UTF-16
representation of the 'Text' value. According to preliminary benchmarks, it is
25% faster than the existing 'encodeUtf8 :: Text -> ByteString' function.

We also support an 'encodeUtf8AsciiEscaped' encoder that allows to special
case encoding of ASCII characters. This is a very useful function for
implementing escaping encoders, e.g., for JSON or HTML output.