Reduce the amount of code generated for string literals
Previously, every Text literal generated a big wad of inefficient
code:

    {-# LANGUAGE OverloadedStrings #-}
    foo :: Text
    foo = "foo"

This would first convert to a String, then to a Text.  To make
matters worse, the code for conversion from String to Text was
inlined at every site where a string literal occurred (expected,
but undesired, behaviour).

In this change, we introduce a direct conversion from Addr# to each
of the Text types, and we ensure that uses of these never result in
excessive code generation.