Improve on previous fix
This version tries to force the real decoding function to be inlined
into each of its callers, which in turn each have different criteria
for backing up a byte. This avoids an extra test at the end of
strict decoding.

While this seems to fix gh-61, I want to beef up the test suite so
that it will correctly detect the bug.