Fix size hints
This fixes a variety of size hint bugs in text's fusion framework. These
issues fell broadly into two classes,

 * Code point/code unit confusion
 * Inappropriate bounds

It seems the most of the latter were introduced when the Size type was
extended to track both upper and lower bounds in f4fc30c. These could
manifest in a variety of issues similar to #197.