Remove rules for takeWhileEnd and dropWhileEnd
it would previously replace `takeWhileEnd p` with

    S.reverse . S.takeWhile p . S.reverseStream

But `S.reverse` can only fuse with its argument, and `S.reverseStream`
only fuses with its consumer. So all possible fusion is within this
pipeline, and in the best case we obtain a direct implementation.

There might be some value in this if we had a
`S.reverseStream/S.reverse` rule, so that in multiple chained invocations
of `takeWhileEnd` we could eliminate some intermediat steps. But no such
rule exists at the moment.