Tidy up extensions and CPP handling
We don't have the means anymore to properly test GHC 6.12 anymore and
code has bitrotten to the point where it makes no sense to even
advertise support for GHC 6.12 anymore.

Starting with this commit `text` now explicitly requires a
Haskell2010+ compiler and the CPP and compiler dependency
specification is updated to reflect this.