Implement liftTyped in the Lift Text instances
`template-haskell-2.16.0.0` adds a `liftTyped` method to `Lift`,
which is like `lift` but returning `Q (TExp t)` instead of `Q Exp`.
`liftTyped` does not have a default implementation in terms of
`lift`, so it must be implemented explicitly in the `Lift Text`
instances that `text` defines. The implementations I chose are based
off of the general approach taken by `Language.Haskell.TH.Syntax`,
such as in this example:
https://gitlab.haskell.org/ghc/ghc/blob/4898df1cc25132dc9e2599d4fa4e1bbc9423cda5/libraries/template-haskell/Language/Haskell/TH/Syntax.hs#L716