Expose UTF-8 validation functions from internal module
This makes it possible for users to validate that a ByteArray is
a well formed UTF-8 sequence. This works both with and without the
SIMDUTF flag, falling back to a pure-Haskell implementation when
SIMDUTF is off.

Pick up a dependency on data-array-byte to be able to refer to
the lifted ByteArray type on older versions of GHC. Use data-array-byte
dep for emulated workflow.