Make Lift Text work under RebindableSyntax
The Lift Int instance produces a polymorphic literal, which desugars to
fromIntegral under RebindableSyntax. Fixing that instance in base may be
too much of a breaking change if a lot of code relies on the accidental
polymorphism. We can at least fix Lift Text by not relying on Lift Int.