Optimize the implementation of Data.Text.concat
Avoid allocating intermediate lists. The algorithm used is not changed,
the total length is calculated in one pass over the input and the output
Text populated in another pass.

Note: In case there is one non-empty Text and one or more empty Texts,
the non-empty Text is copied whereas it would previously be returned as
is. The behavior is unchanged for every other case.